Refactor app setup to support msw in tests
